Fixing Value Conversion Issues

+++ Passions/PASManageArtists.m
@@ -317,7 +317,7 @@
 		
 		// Setup a Timeout
 		void (^timer)(void) = ^{
-			dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, kPASLastFmTimeoutInSec * 0.6 * NSEC_PER_SEC);
+			dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, (NSInteger)(kPASLastFmTimeoutInSec * 0.6 * NSEC_PER_SEC));
 			dispatch_after(popTime, dispatch_get_main_queue(), ^(void){
 				if ([task state] == NSURLSessionTaskStateRunning) {
 					// Task is running too long, cancel it
